reworked the enzy_htp/structure/structure_region submodule
